backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
requests==2.32.3
python-dotenv==1.0.1
websockets==15.0.1
sqlalchemy==2.0.38
alembic==1.14.1
bcrypt==4.2.1
PyJWT==2.10.1
